Break out default route into separate file, add acknowledgements
