+++ delayshell.cc
@@ -7,9 +7,6 @@
 #include <fstream>
 #include <grp.h>
 #include <sys/ioctl.h>
-#include <linux/if.h>
-#include <arpa/inet.h>
-#include <net/route.h>
 
 #include "tundevice.hh"
 #include "exception.hh"
@@ -21,6 +18,7 @@
 #include "address.hh"
 #include "drop_privileges.hh"
 #include "file_descriptor.hh"
+#include "route.hh"
 
 using namespace std;
 
@@ -117,34 +115,7 @@ int main( int argc, char *argv[] )
                 }
 
                 /* Set route the pedantic way */
-                // set route struct to zero
-                struct rtentry route;
-                struct sockaddr_in *addr;
-                int err = 0;
-                memset(&route, 0, sizeof(route));
-
-                // assign the default gateway
-                addr = reinterpret_cast<struct sockaddr_in*>( &route.rt_gateway );
-                addr->sin_family = AF_INET;
-                addr->sin_addr.s_addr = inet_addr(egress_addr.c_str());
-
-                // assign the destination
-                addr = reinterpret_cast<struct sockaddr_in*>( &route.rt_dst );
-                addr->sin_family = AF_INET;
-                addr->sin_addr.s_addr = inet_addr("0.0.0.0");
-
-                // assign the destination mask
-                addr = reinterpret_cast<struct sockaddr_in*>( &route.rt_genmask);
-                addr->sin_family = AF_INET;
-                addr->sin_addr.s_addr = inet_addr("0.0.0.0");
-
-                // set the flags NET/UP
-                route.rt_flags = RTF_UP | RTF_GATEWAY;
-                // make the ioctl
-                if ((err = ioctl(sockfd.num(), SIOCADDRT, &route)) != 0) {
-                   perror("SIOCADDRT failed");
-                   return -1;
-                }
+                Route( move( sockfd ), egress_addr );
 
                 /* create inside listener socket for UDP dns requests */
                 Socket listener_socket_inside( SocketType::UDP );
